Use Binary Dir instead of Source Dir as target for the configured file
We dont want to change things directly in the soure dir (in this case include/internal/). Therefore the destination for adolc_setting.h is changed to CMAKE_BINARY_DIR. To build adolc as dependency the binary file for the destination are generated and the configured file is direcrtly placed there. Adding the create path to include and install allows to build adolc_settings.h.